+++ src/main/java/com/docusign/esign/model/GroupBrands.java
@@ -10,46 +10,21 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 
 /**
- * BrandsResponse
+ * 
  */
+@ApiModel(description = "")
 
-public class BrandsResponse {
-  @JsonProperty("brands")
-  private java.util.List<Brand> brands = null;
-
+public class GroupBrands {
   @JsonProperty("recipientBrandIdDefault")
   private String recipientBrandIdDefault = null;
 
   @JsonProperty("senderBrandIdDefault")
   private String senderBrandIdDefault = null;
 
-  public BrandsResponse brands(java.util.List<Brand> brands) {
-    this.brands = brands;
-    return this;
-  }
+  @JsonProperty("brandOptions")
+  private java.util.List<Brand> brandOptions = null;
 
-  public BrandsResponse addBrandsItem(Brand brandsItem) {
-    if (this.brands == null) {
-      this.brands = new java.util.ArrayList<Brand>();
-    }
-    this.brands.add(brandsItem);
-    return this;
-  }
-
-   /**
-   * The list of brands.
-   * @return brands
-  **/
-  @ApiModelProperty(value = "The list of brands.")
-  public java.util.List<Brand> getBrands() {
-    return brands;
-  }
-
-  public void setBrands(java.util.List<Brand> brands) {
-    this.brands = brands;
-  }
-
-  public BrandsResponse recipientBrandIdDefault(String recipientBrandIdDefault) {
+  public GroupBrands recipientBrandIdDefault(String recipientBrandIdDefault) {
     this.recipientBrandIdDefault = recipientBrandIdDefault;
     return this;
   }
@@ -67,7 +42,7 @@ public class BrandsResponse {
     this.recipientBrandIdDefault = recipientBrandIdDefault;
   }
 
-  public BrandsResponse senderBrandIdDefault(String senderBrandIdDefault) {
+  public GroupBrands senderBrandIdDefault(String senderBrandIdDefault) {
     this.senderBrandIdDefault = senderBrandIdDefault;
     return this;
   }
@@ -85,6 +60,32 @@ public class BrandsResponse {
     this.senderBrandIdDefault = senderBrandIdDefault;
   }
 
+  public GroupBrands brandOptions(java.util.List<Brand> brandOptions) {
+    this.brandOptions = brandOptions;
+    return this;
+  }
+
+  public GroupBrands addBrandOptionsItem(Brand brandOptionsItem) {
+    if (this.brandOptions == null) {
+      this.brandOptions = new java.util.ArrayList<Brand>();
+    }
+    this.brandOptions.add(brandOptionsItem);
+    return this;
+  }
+
+   /**
+   * The list of brands.
+   * @return brandOptions
+  **/
+  @ApiModelProperty(value = "The list of brands.")
+  public java.util.List<Brand> getBrandOptions() {
+    return brandOptions;
+  }
+
+  public void setBrandOptions(java.util.List<Brand> brandOptions) {
+    this.brandOptions = brandOptions;
+  }
+
 
   @Override
   public boolean equals(java.lang.Object o) {
@@ -94,26 +95,26 @@ public class BrandsResponse {
     if (o == null || getClass() != o.getClass()) {
       return false;
     }
-    BrandsResponse brandsResponse = (BrandsResponse) o;
-    return Objects.equals(this.brands, brandsResponse.brands) &&
-        Objects.equals(this.recipientBrandIdDefault, brandsResponse.recipientBrandIdDefault) &&
-        Objects.equals(this.senderBrandIdDefault, brandsResponse.senderBrandIdDefault);
+    GroupBrands groupBrands = (GroupBrands) o;
+    return Objects.equals(this.recipientBrandIdDefault, groupBrands.recipientBrandIdDefault) &&
+        Objects.equals(this.senderBrandIdDefault, groupBrands.senderBrandIdDefault) &&
+        Objects.equals(this.brandOptions, groupBrands.brandOptions);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(brands, recipientBrandIdDefault, senderBrandIdDefault);
+    return Objects.hash(recipientBrandIdDefault, senderBrandIdDefault, brandOptions);
   }
 
 
   @Override
   public String toString() {
     StringBuilder sb = new StringBuilder();
-    sb.append("class BrandsResponse {\n");
+    sb.append("class GroupBrands {\n");
     
-    sb.append("    brands: ").append(toIndentedString(brands)).append("\n");
     sb.append("    recipientBrandIdDefault: ").append(toIndentedString(recipientBrandIdDefault)).append("\n");
     sb.append("    senderBrandIdDefault: ").append(toIndentedString(senderBrandIdDefault)).append("\n");
+    sb.append("    brandOptions: ").append(toIndentedString(brandOptions)).append("\n");
     sb.append("}");
     return sb.toString();
   }